sticky cams: discovered and fixed major flaw in test setup:
We stopped all actors at the end of each runtime.run. The next run restarted all actors with new timers.
This leads to unexpected behaviours when runtime.run is called for short periods of time.

For example:
- The downloader should take 0.5 sec.
- If you let the runtime run for only 0.4 sec, the download in camera_simulation finishes after 0.4.
- When you then run for another 0.2 sec the download starts from the begining.

The solution was to split runtime.run and runtime.advance_time.
The first is now only used once on startup and the latter is used to progress the simulation forward for a given time.