Fix camera removal (#200)
This PR fixes the removal of cameras in the multi_camera_provider.

The problem was that the camera removal works by modifying the dict returned by self.cameras. However, the multi_camera_provider overwrites this method to return a joined dict of all sub-providers. Consequently, the deletion had no effect.

This PR also adds a deprecation warning for the prune_images function which is outdated because images is a dequeue now.