Add MultiClientDisplacementScheduler
This can be seen as a generalization of the LazyWorker with a
configurable number of slots for running coroutines and fair
scheduling between a number of different "clients" where each client has
a LazyWorker-like "waiting"-slot.

Motivation:
On a system with multiple cameras, but one (or multiple) shared
detector(s), images from one camera can repeatedly displace the
detector-requests from another when using LazyWorker (for example
because of differing frame rates). MultiClientDisplacementScheduler
manages access to the shared resource of detectors fairly in that case.
Having multiple running-slots enables better utilization of detector
resources, e.g. in cases where we have multiple detectors running in a
pipeline.