Add /logs page to download `*.log` files using the ui (#360)
### Motivation
When using rosys on a remote device it can be cumbersome to download the
`*.log` files using ssh. To make this easier I want to introduce a
`/logs` page to make the logs downloadable using the ui.
<!-- What problem does this PR solve? Which new feature or improvement
does it implement? -->
<!-- Please provide relevant links to corresponding issues and feature
requests. -->

### Implementation
I choose to take inspirtaion from `video_page.py`. In a similar fashion
we have a list of `*.log` and `*.log.*` files. These can be downloaded
or inspected in the browser. This function could ahev been added to the
`/logging` page. Instead I choose to make a seperate page, in order to
prevent the `/logging` page from getting to complex.
<!-- What is the concept behind the implementation? How does it work?
-->
<!-- Include any important technical decisions or trade-offs made -->

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] I chose meaningful labels (if GitHub allows me to so).
- [x] The implementation is complete.
- [x] Pytests have been added (or are not necessary).
- [x] Documentation has been added (or is not necessary).

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>