Change image representation to pixel array (#356)
### Motivation
Currently, the internal representation of `Image` is data, an encoded
bytes array (usually assumed to be JPEG bytes). This is useful, if we
assume that we get a JPEG from a camera and directly present it as such
to the user, without any further processing. The moment we do any
[processing](https://github.com/zauberzeug/rosys/blob/2f0b816dc7f2b7e2b9b3da2b2806aaa73307b9cb/rosys/vision/image_processing.py#L35),
this advantage falls out of the window and we do more work that we must.

### Implementation

The `data` property no longer exists. Instead, the `array` property can
be used to interact with the image data. `Image.from_jpeg_bytes` can be
used to conveniently construct an image in places where JPEG bytes are
given, and `Image.to_jpeg_bytes` when JPEG bytes are required for
further processing.
Notably, array can no longer be `None` (i.e., we can no longer have
images without any pixel data). I found no problems with that in the
rosys code base, but I'm not sure if there are any use cases in
downstream projects.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] I chose meaningful labels (if GitHub allows me to so).
- [x] The implementation is complete.
- [x] Pytests have been added (or are not necessary).
- [x] Documentation has been added (or is not necessary).

---------

Co-authored-by: Dr. Dennis Wittich <dennis@zauberzeug.com>
Co-authored-by: Niklas Neugebauer <niklas@zauberzeug.com>
Co-authored-by: Falko Schindler <falko@zauberzeug.com>