Install esptool via Lizard's requirements.txt (#375)
### Motivation

A Dependabot alert complained about ecdsa, a dependency of esptool. But
it is strange that RoSys depends on esptool and never uses it directly.

### Implementation

This PR removes esptool from pyproject.toml and extends the installation
of requirements after unpacking the Lizard release artefact. Instead of
installing prompt-toolkit, we now install all requirements.txt.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] I chose meaningful labels (if GitHub allows me to so).
- [x] The implementation is complete.
- [x] Pytests are not necessary.
- [x] Documentation is not necessary.