Lower RtspDevice childprocess pipe size limit to default
This is sufficient because readexact on the stream (which we use)
buffers beyond the limit to fulfill our request anyway. The only effect
of a larger limit is visibile in situations where we cannot keep up with
the frames from there camera where we fail to apply backpressure,
leading to the gstreamer process stuffing the process pipe with old
images.